requirements.txt
python-telegram-bot[job-queue]==20.7
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
# Система записи действий
beautifulsoup4==4.12.2
lxml==4.9.3
# Mini App API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6 